typos
To clarify this kids talk over here:
nsaneforums.com/topic/310042-internet-download-manager-630-build-8

I'm not xanax and the resetter exists since 2015, I only developed the 'GUI' for GitHub and uploaded my source here in order to find problems when IDM decided to change something (yet again) on their product. 

Xanax is not a real developer nor a creator of anything, he steal shit from ru-board repack it and upload it as 'own' 'medicine' on nsane forums. 

The reason I'm not (anymore) on MDL/Nsane or other forums anymore is obvious I'm sick of those bitch fights. 
Waste of energy and lifetime. 

Thanks to